requirements.txt
streamlit
spotipy
pandas
numpy
python-dotenv

